build.rs: propagate Cargo profile.debug to CMake
Propagate the Cargo profile.debug==false setting to the RocksDB build.
RocksDB debug information can be significiant in size (e.g on macOS
librocksdb.a is ~800MiB with default default debuginfo (), or ~200MiB
with debuginfo disabled), but cmake-rs doesn't communicate the Cargo
setting by default so we do it explicitly here.